pdf/pyproject.toml
[tool.poetry]
name = "pdf"
version = "0.1.0"
description = ""
authors = ["Kaito Suzuki <snails8d@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.5.0"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
